varias cosas

+++ demo/src/main/java/repository/RepositoryChef.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface RepositoryChef extends CrudRepository<Chef, Integer> {
+public interface RepositoryChef extends CrudRepository<Chef, Long> {
     List<Client> findByLastName(String lastName);
     List<Client> findBymail(String mail);
 }
